requirements.txt
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
black==19.10b0
Click==7.0
Django==3.0.2
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-camel-case==1.1.2
djangorestframework-jwt==1.11.0
entrypoints==0.3
flake8==3.7.9
mccabe==0.6.1
pathspec==0.7.0
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pytz==2019.3
regex==2020.1.8
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.0
